@@ -1,8 +1,8 @@
-package student.aschm22s.hbrsiCalGenerator.exelImport.service;
+package student.aschm22s.hbrsiCalGenerator.excelImport.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import student.aschm22s.hbrsiCalGenerator.exelImport.domain.HbrsExcelParser;
+import student.aschm22s.hbrsiCalGenerator.excelImport.domain.HbrsExcelParser;
 import student.aschm22s.hbrsiCalGenerator.stundenplanSpecific.veranstaltung.service.VeranstaltungsService;
 
 import java.io.IOException;
