@@ -1,3 +1,3 @@
-export const baseUrl = "https://moin.meister.ovh:8443/api"
+// export const baseUrl = "https://moin.meister.ovh:8443/api"
 //Option for docker-compose
-// export const baseUrl = "http://localhost:8080/api"
+export const baseUrl = "http://localhost:8080/api"
